The four width sites, and the poll literals re-derived from the table
Slice 1d, review finding B4. Four places narrowed or hand-summed a charge in a
way that was safe at epoch 1 and is not at epoch 2, where `DIV r/m32` is raw 492
and `FSQRT` is raw 840.

**B4.1 -- `jit/native_x87.rs`'s per-instruction `raw_clocks: u8 -> u16.** The
widest epoch-1 literal there is `FSQRT`'s 70, which fits a byte; every epoch-2
x87 charge is that count times twelve, so the byte overflows the moment the
metadata is driven from the class table. Widening it now is free -- the field is
read as `u64` one line below -- and `MAX_X87_BLOCK_CORE_CLOCKS` needs no
re-derivation because its compile-time dominance check derives itself from the
metadata table rather than from the field's width, and this commit moves no
metadata value.

**B4.2 -- `StaticAccounting::raw_clocks: u16 -> u32`, with the truncating cast
and the unchecked `+=` gone.** The accumulator was fed by
`slot.kind.raw_clocks() as u16` and summed with a `+=` that panics in debug and
WRAPS in release. `u32` makes overflow unreachable rather than merely checked:
32 slots at the table's widest entry is under 800,000 against four billion, and
`emit` has no refusal path it could take if it were not. `emit_call_out`'s
`static_prefix_raw` widens with it, since it is the same accumulator's value.
`CompiledBlock::raw_clocks` keeps its own `u16` and its existing install-time
refusal, and a test pins that a full block of the widest NATIVE class
(`Idiv32`, 552 raw) still fits it, so the refusal does not become reachable on
ordinary code.

**B4.3 -- `RepLimitPlan::compute`'s `per_iteration` was BUS-ONLY**, so a chunk
could over-admit by 4,096 iterations times the whole core cost of an element.
The per-element class cost now joins it -- UNDER EPOCH 2 ONLY. At epoch 1 that
cost is `StringElem`'s 4 raw, a third of a guest clock at the 1/12 personas and
swamped by the bus terms, and adding it would move `rep_chunk_limit` on every
fixture, which the knob-unset byte-identity merge bar forbids. At epoch 2 it is
48 raw, four whole clocks an element, and the omission is material. The epoch-1
over-admission is recorded at the site as a pre-existing defect that slice 7's
default flip resolves, not one this slice introduces.

**B4.4 -- the four `jit/block.rs` poll literals** (17, 21, 28 and
`MEMORY_POLL_RAW_CORE_CLOCKS` = 5) are re-derived from the running table. A
skipped poll iteration must charge what a really-run one charges or the
projection lane and the interpreter disagree about how much guest time an elided
span consumed; under epoch 2 a skipped `IN AL,DX` poll would have cost 17 raw
where a run one costs 168. Each derivation names the exact instruction sequence
its shape matches -- the structural predicates above them are what guarantee the
sequence -- and the four hand-summed literals survive as COMPILE-TIME tripwires:
a `const _: () = { ... }` block asserts the derivations reproduce 17 / 21 / 28 /
5 exactly against the epoch-1 column, so a table edit that would desynchronise
the projection fails the build.

`EPOCH1` gains a `const` twin (`EPOCH1_CONST`) so those assertions can evaluate:
a `const` may not read a `static`, and `EPOCH1` must stay a `static` for its
address to be stable.

Epoch 1 is unchanged: 2401 izarravm-cpu tests and the whole izarravm-machine
suite pass.

+++ crates/izarravm-cpu/src/jit/native_x87.rs
@@ -330,7 +330,16 @@ pub(crate) struct NativeX87MemoryAccess {
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) struct NativeX87Metadata {
-    pub(crate) raw_clocks: u8,
+    /// The instruction's raw core clocks before `fp_timing_class`'s scaling.
+    ///
+    /// `u16`, not `u8`, since slice 1d (review B4.1). The widest epoch-1 literal
+    /// here is 70 (`FSQRT`), which fits a byte; every epoch-2 x87 charge is that
+    /// count times twelve -- `FSQRT` alone is raw 840 -- so the byte overflows
+    /// the moment this table is driven from the class table. Widening it ahead of
+    /// that is free: the field is read as `u64` one line below, and
+    /// `MAX_X87_BLOCK_CORE_CLOCKS`'s compile-time dominance check re-derives
+    /// itself from this table rather than from the field's width.
+    pub(crate) raw_clocks: u16,
     pub(crate) fp_class: FpOpClass,
     pub(crate) memory: Option<NativeX87MemoryAccess>,
     pub(crate) pops: bool,
